+++ test/quantization/eager/test_equalize_eager.py
@@ -2,7 +2,7 @@ import torch
 import torch.nn as nn
 
 from torch.testing._internal.common_quantization import QuantizationTestCase
-from torch.quantization.fuse_modules import fuse_modules
+from torch.ao.quantization.fuse_modules import fuse_modules
 
 import torch.quantization._equalize as _equalize
